More changes to appease the Pull Request.
