+++ src/Core/Room.cs
@@ -7,7 +7,7 @@ public class Room {
     static object RoomsListLock = new object();
     protected static Dictionary<string, Room> rooms = new();
 
-    protected List<Client> clients = new();
+    List<Client> clients = new();
     protected object roomLock = new object();
 
     public int Id { get; private set; }
@@ -82,10 +82,8 @@ public class Room {
     
     public static Room GetOrAdd(string name, bool autoRemove = false) {
         lock(RoomsListLock) {
-            if (!Room.Exists(name)) {
-                if (Configuration.ServerConfiguration.AmbassadorRooms.Contains(name)) return new AmbassadorRoom(name);
+            if (!Room.Exists(name)) 
                 return new Room(name, autoRemove: autoRemove);
-            }
             return rooms[name];
         }
     }
@@ -113,10 +111,7 @@ public class Room {
         roomInfo.Add(false); // is password protected
         roomInfo.Add((short)clients.Count); // player count
         roomInfo.Add((short)4096); // max player count
-        NetworkArray roomVars = new();
-        AddRoomData(roomVars);
-        for (int i=0;i<RoomVariables.Length;i++) roomVars.Add(RoomVariables[i]);
-        roomInfo.Add(roomVars); // variables (plus added data)
+        roomInfo.Add(GetRoomVars()); // variables (plus added data)
         roomInfo.Add((short)0); // spectator count
         roomInfo.Add((short)0); // max spectator count
 
@@ -148,9 +143,7 @@ public class Room {
         r1.Add(false);
         r1.Add((short)clients.Count); // player count
         r1.Add((short)4096); // max player count
-        NetworkArray vars = new();
-        AddRoomData(vars);
-        r1.Add(vars);
+        r1.Add(GetRoomVars());
         r1.Add((short)0);
         r1.Add((short)0);
 
@@ -161,146 +154,5 @@ public class Room {
         return NetworkObject.WrapObject(0, 15, obj).Serialize();
     }
 
-    /// <summary>
-    /// Add extra data for the client to recieve upon joining the room.
-    /// </summary>
-    /// <param name="vars">Existing room variables.</param>
-    internal virtual void AddRoomData(NetworkArray vars) {}
-
-    private int alertId = -1;
-    private Random random = new Random();
-
-    List<AlertInfo> alerts = new();
-
-    public void AddAlert(AlertInfo alert) {
-        alerts.Add(alert);
-        ResetAlertTimer(alert);
-    }
-
-    public void SendAllAlerts(Client client) {
-        return; // Disables joining ongoing alerts (since it doesn't work properly).
-        
-        foreach (AlertInfo alert in alerts) {
-            if (alert.IsRunning()) StartAlert(alert, client);
-        }
-    }
-
-
-    private void StartAlert(AlertInfo alert, Client? specificClient = null) {
-        NetworkArray NewRoomVariables = new();
-        NewRoomVariables.Add(NetworkArray.VlElement(REDALERT_START, alertId++, isPersistent: true));
-        NewRoomVariables.Add(NetworkArray.VlElement(REDALERT_TYPE, alert.type, isPersistent: true));
-        double duration = (alert.endTime - DateTime.Now).TotalSeconds;
-        NewRoomVariables.Add(NetworkArray.VlElement(REDALERT_LENGTH, alert.type == "1" ? alert.redAlertDuration : duration, isPersistent: true));
-        if (alert.type == "1") {
-            NewRoomVariables.Add(NetworkArray.VlElement(REDALERT_TIMEOUT, duration, isPersistent: true));
-        } else if (alert.type == "3") {
-            alert.songId = random.Next(0, alert.songs);
-            NewRoomVariables.Add(NetworkArray.VlElement(REDALERT_SONG, (double)alert.songId, isPersistent: true));
-        }
-        NetworkPacket packet = Utils.VlNetworkPacket(NewRoomVariables, Id);
-        if (specificClient is null) {
-            RoomVariables = NewRoomVariables;
-            Send(packet);
-            Console.WriteLine("Started event " +alert + " in room " + Name);
-        } else {
-            specificClient.Send(packet);
-            Console.WriteLine("Added " + specificClient.PlayerData.DiplayName + " to event " + alert + " with " + duration + " seconds remaining");
-        }
-    }
-
-    void ResetAlertTimer(AlertInfo alert) {
-        System.Timers.Timer? timer = alert.timer;
-        if (timer != null) {
-            timer.Stop();
-            timer.Close();
-        }
-        DateTime startTime = DateTime.Now.AddMilliseconds(random.Next(alert.minTime * 1000, alert.maxTime * 1000));
-        DateTime endTime = startTime.AddSeconds(alert.duration);
-        for (int i = 0; i < alerts.IndexOf(alert); i++) {
-            // Prevent overlap between two events.
-            if (alerts[i].Overlaps(endTime)) {
-                startTime = alerts[i].endTime.AddSeconds(5);
-                endTime = startTime.AddSeconds(alert.duration);
-            }
-        }
-        timer = new System.Timers.Timer((startTime - DateTime.Now).TotalMilliseconds);
-        timer.AutoReset = false;
-        timer.Enabled = true;
-        timer.Elapsed += (sender, e) => StartAlert(alert);
-        timer.Elapsed += (sender, e) => ResetAlertTimer(alert);
-        alert.timer = timer;
-        Console.WriteLine("Event " + alert + " in " + Name + " scheduled for " + startTime.ToString("MM/dd/yyyy HH:mm:ss tt") + " (in " + (startTime - DateTime.Now).TotalSeconds + " seconds)");
-        alert.startTime = startTime;
-        alert.endTime = endTime;
-    }
-
-    private const string REDALERT_START = "RA_S";
-    private const string REDALERT_TYPE = "RA_A";
-    private const string REDALERT_LENGTH = "RA_L";
-    private const string REDALERT_TIMEOUT = "RA_T";
-    private const string REDALERT_SONG = "RA_SO";
-
-    public class AlertInfo {
-        public readonly string type;
-        public readonly double duration;
-        public readonly int minTime;
-        public readonly int maxTime;
-        public readonly int redAlertDuration;
-        public readonly int songs;
-        public int songId;
-
-        public DateTime startTime {
-            get {
-                return newStartTime;
-            }
-            set {
-                oldStartTime = newStartTime;
-                newStartTime = value;
-            }
-        }
-
-        public DateTime endTime {
-            get {
-                return newEndTime;
-            }
-            set {
-                oldEndTime = newEndTime;
-                newEndTime = value;
-            }
-        }
-
-        private DateTime newStartTime;
-        private DateTime newEndTime;
-        private DateTime oldStartTime;
-        private DateTime oldEndTime;
-
-        public System.Timers.Timer? timer = null;
-
-        public AlertInfo(string type, double duration = 20.0, int minTime = 30, int maxTime = 240, int redAlertDuration = 60, int songs = 16) {
-            this.type = type;
-            this.duration = duration;
-            this.minTime = minTime;
-            this.maxTime = maxTime;
-            this.redAlertDuration = redAlertDuration;
-            this.songs = songs;
-        }
-
-        public bool Overlaps(DateTime time) {
-            return (time >= oldStartTime && time <= oldEndTime);
-        }
-
-        public bool IsRunning() {
-            return Overlaps(DateTime.Now);
-        }
-
-        public override string ToString() {
-            return type switch {
-                "1" => "RedAlert",
-                "2" => "DiscoAlert",
-                "3" => "DanceOff",
-                _ => type
-            };
-        }
-    }
+    internal virtual NetworkArray GetRoomVars() { return RoomVariables; }
 }
